+++ Assets/Scripts/NPC/NPCGoap.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 69429fc0a03dccc42a4dd8c0a2ce2009
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
